hook for cohort dashboard
